+++ api/src/main/java/ca/bc/gov/educ/api/course/EducCourseApiApplication.java
@@ -1,26 +1,12 @@
 package ca.bc.gov.educ.api.course;
 
-import ca.bc.gov.educ.api.course.model.dto.StudentCourse;
-import ca.bc.gov.educ.api.course.model.entity.StudentCourseEntity;
-import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.context.SecurityContextHolder;
-
-import ca.bc.gov.educ.api.course.model.dto.Course;
-import ca.bc.gov.educ.api.course.model.entity.CourseEntity;
 
 @SpringBootApplication
-@EnableGlobalMethodSecurity(prePostEnabled = true)
 @EnableCaching
 public class EducCourseApiApplication {
 
@@ -32,43 +18,4 @@ public class EducCourseApiApplication {
 		logger.debug("########Started API");
 	}
 
-	@Bean
-	public ModelMapper modelMapper() {
-
-		ModelMapper modelMapper = new ModelMapper();
-		modelMapper.typeMap(CourseEntity.class, Course.class);
-		modelMapper.typeMap(Course.class, CourseEntity.class);
-
-		modelMapper.typeMap(StudentCourseEntity.class, StudentCourse.class);
-		modelMapper.typeMap(StudentCourse.class, StudentCourseEntity.class);
-		return modelMapper;
-	}
-	
-	@Configuration
-	static
-	class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
-	  /**
-	   * Instantiates a new Web security configuration.
-	   * This makes sure that security context is propagated to async threads as well.
-	   */
-	  public WebSecurityConfiguration() {
-	    super();
-	    SecurityContextHolder.setStrategyName(SecurityContextHolder.MODE_INHERITABLETHREADLOCAL);
-	  }
-	  @Override
-	  public void configure(WebSecurity web) {
-		  web.ignoring().antMatchers("/api/v1/api-docs-ui.html",
-				  "/api/v1/swagger-ui/**", "/api/v1/api-docs/**",
-				  "/actuator/health","/actuator/prometheus", "/health");
-	  }
-		@Override
-		protected void configure(final HttpSecurity http) throws Exception {
-			http.authorizeRequests()
-					.anyRequest()
-					.authenticated()
-					.and()
-					.oauth2ResourceServer()
-					.jwt();
-		}
-	}
 }
